pyproject.toml
[tool.poetry]
name = "decluttergcal"
version = "0.1.0"
description = "CLI to batch delete Google Calendar events"
authors = ["Tendai Charura <Tendai.Charura@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.6.1"
google-api-python-client = "^2.58"
google-auth-oauthlib = "^0.5.2"

[tool.poetry.scripts]
gcal = 'decluttergcal.cli.cli:app'

[tool.poetry.dev-dependencies]
black = "^22.6.0"
mkdocs-material = "^8.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"